cs_module/requirements.txt
requests
numpy
flask
scipy
psycopg2-binary
uuid
python-dateutil